requirements.txt
psycopg2~=2.9.9
FlightRadarAPI~=1.3.31
pandas~=2.2.2
marshmallow~=3.21.3
flask~=3.0.3
Flask-RESTful~=0.3.10
requests~=2.32.3
setuptools~=68.2.0
python-dateutil~=2.9.0.post0
streamlit~=1.36.0
APScheduler~=3.10.4
SQLAlchemy~=2.0.31